libs/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "libs"
version = "0.1.0"
description = "Shared utilities for ai-content-farm (internal)"
authors = [ { name = "ai-content-farm" } ]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity",
    "azure-storage-blob",
    "azure-storage-queue",
    "azure-servicebus",
    "azure-keyvault-secrets",
    "azure-monitor-opentelemetry~=1.6.4",  # Application Insights integration
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pydantic",
    "pydantic-settings",
    "aiofiles",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "requests",
    "tenacity",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
py-modules = ["shared_models", "blob_storage", "storage_queue_client"]
